+++ src/preprocessor/handlers/code_handler/plot_obj_parser/ohlc_parser.rs
@@ -1,17 +1,17 @@
 use super::until::{Map, must_translate};
-use crate::translate;
-use anyhow::{Result, anyhow};
+use crate::{translate, translate_enum};
+use anyhow::Result;
 use plotly::Ohlc;
 
 pub fn parse_ohlc_data(
     ohlc_obj: &mut serde_json::Value,
     map: &Map,
 ) -> Result<Box<Ohlc<String, f64>>> {
-    let x: Vec<String> = must_translate(ohlc_obj, "x")?;
-    let open: Vec<f64> = must_translate(ohlc_obj, "open")?;
-    let high: Vec<f64> = must_translate(ohlc_obj, "high")?;
-    let low: Vec<f64> = must_translate(ohlc_obj, "low")?;
-    let close: Vec<f64> = must_translate(ohlc_obj, "close")?;
+    let x: Vec<String> = must_translate(ohlc_obj, map, "x")?;
+    let open: Vec<f64> = must_translate(ohlc_obj, map, "open")?;
+    let high: Vec<f64> = must_translate(ohlc_obj, map, "high")?;
+    let low: Vec<f64> = must_translate(ohlc_obj, map, "low")?;
+    let close: Vec<f64> = must_translate(ohlc_obj, map, "close")?;
     let ohlc = Ohlc::new(x, open, high, low, close);
     let ohlc = translate! {
         *ohlc,
@@ -26,20 +26,18 @@ pub fn parse_ohlc_data(
         (hover_text_array, Vec<String>),
         (tick_width, f64),
     }?;
-    let ohlc = if let Some(visible) = ohlc_obj.get_mut("visible")
-        && visible.is_string()
-    {
-        use plotly::common::Visible;
-        let visible = match visible.as_str().unwrap_or_else(|| unreachable!()) {
-            "true" => Visible::True,
-            "false" => Visible::False,
+
+    use plotly::common::Visible;
+    let ohlc = translate_enum! {
+        ohlc,
+        ohlc_obj,
+        map,
+        (visible, {
+            "true" =>       Visible::True,
+            "false" =>      Visible::False,
             "legendonly" => Visible::LegendOnly,
-            unexpected => return Err(anyhow!("{unexpected} can't be visible")),
-        };
-        ohlc.visible(visible)
-    } else {
-        ohlc
-    };
+        }),
+    }?;
 
     // UNEXPECTED: The other methods of the `Ohlc` return only `self`, not boxed `self`.
     Ok(Box::new(ohlc))