+++ src/preprocessor/handlers/code_handler/plot_obj_parser/scatter_geo_parser.rs
@@ -1,14 +1,14 @@
 use super::until::{Map, must_translate};
-use crate::translate;
-use anyhow::{Result, anyhow};
+use crate::{translate, translate_enum};
+use anyhow::Result;
 use plotly::ScatterGeo;
 
 pub fn parse_scatter_geo_data(
     sg_obj: &mut serde_json::Value,
     map: &Map,
 ) -> Result<Box<ScatterGeo<f64, f64>>> {
-    let lat: Vec<f64> = must_translate(sg_obj, "lat")?;
-    let lon: Vec<f64> = must_translate(sg_obj, "lon")?;
+    let lat: Vec<f64> = must_translate(sg_obj, map, "lat")?;
+    let lon: Vec<f64> = must_translate(sg_obj, map, "lon")?;
     let sg = ScatterGeo::new(lat, lon);
     let sg = translate! {
         sg,
@@ -32,24 +32,23 @@ pub fn parse_scatter_geo_data(
         (subplot, String),
         (below, String),
     }?;
-    let sg = if let Some(mode) = sg_obj.get_mut("mode")
-        && mode.is_string()
-    {
-        use plotly::common::Mode;
-        let mode = match mode.as_str().unwrap_or_else(|| unreachable!()) {
-            "lines" => Mode::Lines,
-            "markers" => Mode::Markers,
-            "text" => Mode::Text,
-            "linesmarkers" => Mode::LinesMarkers,
-            "linestext" => Mode::LinesText,
-            "markerstext" => Mode::MarkersText,
-            "linemarkerstext" => Mode::LinesMarkersText,
-            "none" => Mode::None,
-            unexpected => return Err(anyhow!("{unexpected} can't be mode")),
-        };
-        sg.mode(mode)
-    } else {
-        sg
-    };
+
+    use plotly::common::Mode;
+    let sg = translate_enum! {
+        sg,
+        sg_obj,
+        map,
+        (mode, {
+            "lines" =>          Mode::Lines,
+            "markers" =>        Mode::Markers,
+            "text" =>           Mode::Text,
+            "linesmarkers" =>   Mode::LinesMarkers,
+            "linestext" =>      Mode::LinesText,
+            "markerstext" =>    Mode::MarkersText,
+            "linemarkerstext" =>Mode::LinesMarkersText,
+            "none" =>           Mode::None,
+        }),
+    }?;
+
     Ok(sg)
 }