+++ src/preprocessor/handlers/code_handler/plot_obj_parser/scatter_mapbox_parser.rs
@@ -1,14 +1,14 @@
 use super::until::{Map, must_translate};
-use crate::translate;
-use anyhow::{Result, anyhow};
+use crate::{translate, translate_enum};
+use anyhow::Result;
 use plotly::ScatterMapbox;
 
 pub fn parse_scatter_mapbox_data(
     sm_obj: &mut serde_json::Value,
     map: &Map,
 ) -> Result<Box<ScatterMapbox<f64, f64>>> {
-    let lat: Vec<f64> = must_translate(sm_obj, "lat")?;
-    let lon: Vec<f64> = must_translate(sm_obj, "lon")?;
+    let lat: Vec<f64> = must_translate(sm_obj, map, "lat")?;
+    let lon: Vec<f64> = must_translate(sm_obj, map, "lon")?;
     let sm = ScatterMapbox::new(lat, lon);
     let sm = translate! {
         sm,
@@ -33,24 +33,23 @@ pub fn parse_scatter_mapbox_data(
         (below, String),
         (meta, String),
     }?;
-    let sm = if let Some(mode) = sm_obj.get_mut("mode")
-        && mode.is_string()
-    {
-        use plotly::common::Mode;
-        let mode = match mode.as_str().unwrap_or_else(|| unreachable!()) {
-            "lines" => Mode::Lines,
-            "markers" => Mode::Markers,
-            "text" => Mode::Text,
-            "linesmarkers" => Mode::LinesMarkers,
-            "linestext" => Mode::LinesText,
-            "markerstext" => Mode::MarkersText,
-            "linemarkerstext" => Mode::LinesMarkersText,
-            "none" => Mode::None,
-            unexpected => return Err(anyhow!("{unexpected} can't be mode")),
-        };
-        sm.mode(mode)
-    } else {
-        sm
-    };
+
+    use plotly::common::Mode;
+    let sm = translate_enum! {
+        sm,
+        sm_obj,
+        map,
+        (mode, {
+            "lines" =>          Mode::Lines,
+            "markers" =>        Mode::Markers,
+            "text" =>           Mode::Text,
+            "linesmarkers" =>   Mode::LinesMarkers,
+            "linestext" =>      Mode::LinesText,
+            "markerstext" =>    Mode::MarkersText,
+            "linemarkerstext" =>Mode::LinesMarkersText,
+            "none" =>           Mode::None,
+        }),
+    }?;
+
     Ok(sm)
 }